+++ CMakeLists.txt
@@ -110,8 +110,8 @@ set(UMAPANALYSIS_SETTINGS
 
 set(UMAPANALYSIS_DISTANCES
     src/hnsw/space_corr.h
+    src/knncolle_base/Matrix_parallel.h
     src/knncolle_hnsw/knncolle_hnsw_parallel.h
-    src/knncolle/Matrix_parallel.h
 )
 
 set(UMAPANALYSIS_AUX
