+++ src/knncolle_base/Matrix_parallel.h
