+++ src/knncolle_hnsw/knncolle_hnsw_parallel.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <knncolle_hnsw/knncolle_hnsw.hpp>
-#include "knncolle/Matrix_parallel.h"
+#include "knncolle_base/Matrix_parallel.h"
 
 #include <atomic>
 #include <thread>
